disable passing of hostname to bearssl
in principle we want this for SNI, but some real mail servers have
certificates with wrong name, and further work would be needed to make
bearssl just accept it. the vtable hook trick done to make TLSA
records override the lack of a trust anchor doesn't work, because
bearssl doesn't even identify which certificate in the chain is for
the end entity without a matching hostname.