factor public key hashing out of dane checking
this commit introduces a pkey_hash function/file that takes a BearSSL
br_x509_pkey structure and produces sha256 and sha512 hashes for the
DER encoded form of the key, and modifies the dane check_tlsa function
to use the resulting hashes as input in place of the key. along with
this, support for byte-by-byte comparison of public key is removed. as
was already done for certificates, tlsa records that specify matching
type 0 (byte-by-byte comparison) for public keys are rewritten as
equivalent rules for the sha256 hash of the key.