handle domains with broken auth ns that can't give negative TLSA answer
DNS spec has always been clear about how to handle queries for unknown
RR types, but some broken authoritative nameservers produce no
response or an error when attempting to lookup TLSA records. a
DNSSEC-validating recursive/caching nameserver must pass this
condition on as a failure to resolve; it can't synthesize a negative
answer.

in order not to make it impossible to deliver to domains with broken
auth nameservers, when TLSA lookup fails, perform a second check to
determine if the zone is signed. a CNAME query is used for this
because it won't follow CNAMEs and because it's universally
compatible. if the result succeeds and AD=0, this means our nameserver
has validated that the zone is insecure, and can be treated as not
having TLSA records. otherwise, the zone is actually or potentially
authenticated and we must treat the failure to lookup TLSA as a temp
failure.

special thanks to Viktor Dukhovni for explaining and reminding me of
this failure mode, which I'd forgotten about in the previous fix
(commit c3d47f4d207817928414a9ec6a277eed01332564) for mishandling of
insecure zones. hopefully I've gotten it right now.