factor test for insecure zone into its own function, use A+CNAME
ANY queries are deprecated and should not be used. since some broken
servers give NotImp for CNAME, first perform A query. but if the
answer for the A query does not have AD bit set, that's not sufficient
to determine that the zone was insecure. it's possible it was a CNAME
in an authenticated zone pointing to an insecure zone. if the answer
for the A query with AD bit clear contains a CNAME RR, perform the
CNAME query to determine if the original zone was insecure.

the new is_insecure function returns true only if the zone is proven
insecure. a result of false means either authenticated or
inconclusive.