fix handling of recipient domains with cname but no mx
the presence of a cname rr in the answer prevented the code path for
fallback to no-mx case from being taken, since in this case the answer
count is nonzero.

instead, count the mx records in the answer first, and fallback if
that count is zero.

bail out early on any error (e.g. servfail). while, as long as the
nameserver is trustworthy, it would work to keep the check with the
fallback path like before, it makes more sense not to even look at the
answer in this case.